MOD: reconstruct test code
